interim commit. more flexible value references.
This is to differentiate between values that need to be garbage
collected, and values that can be copied.

Aliases V to Vu.
Moves the ref smart ptr from V to variants in Vu.
Use tuple structs for D1/D2.
Add tracing functions for new refcounted structs.
Modifies provide to not wrap functions in ref ptrs.
Changes array shape to vec of usize.